pyproject.toml
[tool.poetry]
name = "tx-generator"
version = "0.1.0"
description = ""
authors = ["clement.l <maple3312936@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.5"
langchain = "^0.3.0"
langchain-chroma = "^0.1.4"
langchain-community = "^0.3.0"
langchain-anthropic = "^0.2.0"
langchain-openai = "^0.2.0"
langchain-google-vertexai = "^2.0.0"
langgraph = "^0.2.21"
web3 = "^7.2.0"
jq = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
